% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicottomoose.R
\name{dicottomoose}
\alias{dicottomoose}
\title{A dicottomoose object see https://www.fishbase.se/keys/allkeys.php for a list of keys}
\value{
a dicottomoose object
}
\description{
A dicottomoose object see https://www.fishbase.se/keys/allkeys.php for a list of keys

A dicottomoose object see https://www.fishbase.se/keys/allkeys.php for a list of keys
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{dicottomoose$new()}}
\item \href{#method-print}{\code{dicottomoose$print()}}
\item \href{#method-toDataTree}{\code{dicottomoose$toDataTree()}}
\item \href{#method-toPolyclave}{\code{dicottomoose$toPolyclave()}}
\item \href{#method-clone}{\code{dicottomoose$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dicottomoose$new(df, desc = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{a dataframe}

\item{\code{desc}}{a description}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dicottomoose$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toDataTree"></a>}}
\subsection{Method \code{toDataTree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dicottomoose$toDataTree(
  includeStatementNodes = FALSE,
  includeLoneLeafNodes = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toPolyclave"></a>}}
\subsection{Method \code{toPolyclave()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dicottomoose$toPolyclave(delim = ";")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dicottomoose$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
