% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importVibrio}
\alias{importVibrio}
\title{Load consensus matrix from I. Noguerola and A.R. Blanch "Identification of Vibrio spp. with a set of dichotomous keys" doi:10.1111/j.1365-2672.2008.03730.x}
\usage{
importVibrio(consensus)
}
\description{
Load consensus matrix from I. Noguerola and A.R. Blanch "Identification of Vibrio spp. with a set of dichotomous keys" doi:10.1111/j.1365-2672.2008.03730.x
}
