% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{looKey}
\alias{looKey}
\title{Leave-one-out evaluation of a key builder}
\usage{
looKey(images, build, score, groups = NULL)
}
\arguments{
\item{images}{[imageSet()] data frame of all images.}

\item{build}{Function of the training row indices returning a key.}

\item{score}{Function of (key, test row index) returning a one-row scores
matrix for that image.}

\item{groups}{As in [evaluateKey()].}
}
\value{
As [evaluateKey()].
}
\description{
Leave-one-out evaluation of a key builder
}
