% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision-python.R
\name{modelMeta}
\alias{modelMeta}
\title{Model details as metadata rows}
\usage{
modelMeta(model)
}
\arguments{
\item{model}{A [visionModel()].}
}
\value{
A key/value data frame with `model_*` keys.
}
\description{
Model details as metadata rows
}
