% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision-python.R
\name{embedTexts}
\alias{embedTexts}
\title{Embed texts}
\usage{
embedTexts(model, texts)
}
\arguments{
\item{model}{A \code{\link[=visionModel]{visionModel()}}.}

\item{texts}{Character vector.}
}
\value{
A \verb{texts x d} matrix with the texts as row names.
}
\description{
Embed texts
}
