% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportWizard}
\alias{exportWizard}
\title{Export an interactive dichotomous tree
see https://www.fishbase.se/keys/allkeys.php for a list of keys}
\usage{
exportWizard(moose, displayLinks = TRUE, order = "parsimony")
}
\arguments{
\item{moose}{a moose object}

\item{displayLinks}{display HTML anchor links for navigation}

\item{order}{order of statements, 'original' if imported from an existing tree, 'parsimony' for maximum parsimony}
}
\description{
Export an interactive dichotomous tree
see https://www.fishbase.se/keys/allkeys.php for a list of keys
}
