% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importFishbase}
\alias{importFishbase}
\title{Import a dichotomous key from fishbase
see https://www.fishbase.se/keys/allkeys.php for a list of keys}
\usage{
importFishbase(
  keycode,
  fishbaseUrl = "https://www.fishbase.se/",
  separateTerms = TRUE
)
}
\arguments{
\item{keycode}{representing some taxa on fishbase}

\item{fishbaseURL}{fishbase url (swedish one seems best?)}
}
\value{
a dicottomoose object
}
\description{
Import a dichotomous key from fishbase
see https://www.fishbase.se/keys/allkeys.php for a list of keys
}
