% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moose-package.R
\docType{package}
\name{moose-package}
\alias{moose-package}
\title{moose: manage, manipulate, and mine dichotomous keys}
\description{
Tools to import dichotomous identification keys (currently from FishBase),
represent them as tables of couplets and leads, and convert them to other
tree representations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leipzig/moose}
  \item Report bugs at \url{https://github.com/leipzig/moose/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Leipzig \email{leipzig@gmail.com} (\href{https://orcid.org/0000-0001-7224-9620}{ORCID})

Authors:
\itemize{
  \item Jeremy Leipzig \email{leipzig@gmail.com} (\href{https://orcid.org/0000-0001-7224-9620}{ORCID})
}

}
\keyword{internal}
