% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moose.R
\name{moose}
\alias{moose}
\title{A moose object see https://www.fishbase.se/keys/allkeys.php for a list of keys}
\value{
a moose object
}
\description{
A moose object see https://www.fishbase.se/keys/allkeys.php for a list of keys

A moose object see https://www.fishbase.se/keys/allkeys.php for a list of keys
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{moose$new()}}
\item \href{#method-print}{\code{moose$print()}}
\item \href{#method-toDataTree}{\code{moose$toDataTree()}}
\item \href{#method-toPolyclave}{\code{moose$toPolyclave()}}
\item \href{#method-clone}{\code{moose$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{moose$new(df, desc = NA, meta = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{a dataframe}

\item{\code{desc}}{a description}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{moose$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toDataTree"></a>}}
\subsection{Method \code{toDataTree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{moose$toDataTree(includeStatementNodes = FALSE, includeLoneLeafNodes = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toPolyclave"></a>}}
\subsection{Method \code{toPolyclave()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{moose$toPolyclave(delim = ";")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{moose$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
