% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moose.R
\name{rawhtml}
\alias{rawhtml}
\title{A raw HTML object with the potential to be a moose}
\value{
a rawhtml object
}
\description{
A raw HTML object with the potential to be a moose

A raw HTML object with the potential to be a moose
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{rawhtml$print()}}
\item \href{#method-new}{\code{rawhtml$new()}}
\item \href{#method-clone}{\code{rawhtml$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rawhtml$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rawhtml$new(
  keycode,
  fishbaseUrl = "https://www.fishbase.se/",
  separateTerms = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rawhtml$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
