% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibrio-data.R
\docType{data}
\name{vibrio}
\alias{vibrio}
\title{Vibrio consensus table form}
\format{A data frame}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(vibrio)
}
\description{
Consensus matrix from
}
\examples{
data(vibrio)
importVibrio(vibrio)
\donttest{iplotCurves(phe, times)}
}
\references{
I. Noguerola and A.R. Blanch "Identification of Vibrio spp. with a set of dichotomous keys"
(\href{https://doi.org/10.1111/j.1365-2672.2008.03730.x}{PubMed})
}
\keyword{datasets}
